+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-## Version 1.0
+## Version 1.1.0
+2022-08-16
+- Localizations for Français and Español.
+- List of languages in the project preferences makes use of enum-resolve so language names
+  can appear in the user's language, in the alphabetical order for that language.
+- Small code improvements.
 
+## Version 1.0.0
+2022-07-15
 Première version.
 Versión inicial.
 الإصدار الأولي.